Hilt and MVVM added
