--- app/src/main/java/com/marvelapp/MainActivity.kt
@@ -1,23 +1,0 @@
-package com.marvelapp
-
-import android.os.Bundle
-import android.view.animation.AnimationUtils
-import androidx.appcompat.app.AppCompatActivity
-import com.marvelapp.databinding.ActivityMainBinding
-
-class MainActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityMainBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-        binding.marvelLogoSplashImageView.startAnimation(
-            AnimationUtils.loadAnimation(
-                this,
-                R.anim.vertical_slide
-            )
-        )
-    }
-}
