+++ app/src/test/java/com/marvelapp/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
 package com.marvelapp
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
