+++ settings.gradle
@@ -1,2 +1,3 @@
 include ':app'
 rootProject.name = "MarvelApp"
+include ':data'
